"last of v5dev working."

+++ scripts/restore.sh
@@ -45,7 +45,7 @@ SECONDS=0
 # $bk is the backed up site.
 
 # Set script name for general file use
-scriptname='pleasy-restore'
+scriptname='restore'
 verbose="none"
 
 # Help menu
@@ -141,6 +141,8 @@ else
   bk=$1
 fi
 
+echo "Restoring site $bk to $sitename_var"
+
 if [[ "$bk" == prod ]] && [[ ! "$prod_method" == "git" ]]; then
   echo "Sorry not able to handle restoring prod unless it is method git."
   exit 0
@@ -156,7 +158,7 @@ if [[ "$bk" == "prod" ]] && [[ "$prod_method" == "git" ]] && [[ "$sitename_var" 
   ssh $prod_alias -t "./restoreprod.sh $prod_docroot $prod_gitrepo"
   # The script does it all. No need for anything else.
   exit 0
-  elif [[ "$bk" == "prod" ]] && [[ "$prod_method" == "git" ]] ; then
+elif [[ "$bk" == "prod" ]] && [[ "$prod_method" == "git" ]]; then
   echo "Restoring production site to site: $sitename_var"
   # First get the database
   #scp "$prod_alias:proddb/prod.sql" "$folderpath/sitebackups/prod/$Bname.sql"
@@ -164,37 +166,46 @@ if [[ "$bk" == "prod" ]] && [[ "$prod_method" == "git" ]] && [[ "$sitename_var" 
 
   # Check if database is already present
   if [[ -f "$folderpath/sitebackups/proddb/prod.sql" ]]; then
-  ocmsg "Pull the database down to proddb." debug
-  cd $folderpath/sitebackups/proddb/
-  git fetch --all
-  #git checkout -b backup-master
-  git reset --hard origin/master
+    if [[ "$(git config --get remote.origin.url)" == "$prod_gitdb" ]]; then
+      ocmsg "Pull the database down to proddb." debug
+      cd $folderpath/sitebackups/proddb/
+      git fetch --all
+      #git checkout -b backup-master
+      git reset --hard origin/master
+    else
+      removedb="yes"
+    fi
   else
-  # Check if proddb exits
-  if [[ -d "$folderpath/sitebackups/proddb/" ]]; then
-    rm "$folderpath/sitebackups/proddb" -rf
+    removedb="yes"
   fi
-  git clone $gitdb "$folderpath/sitebackups/proddb"
+  if [[ "$removedb" == "yes" ]]; then
+    # Check if proddb exits
+    if [[ -d "$folderpath/sitebackups/proddb/" ]]; then
+      echo "removing proddb"
+      rm "$folderpath/sitebackups/proddb" -rf
+    fi
+    echo "Cloning $prod_gitdb into $folderpath/sitebackups/proddb"
+    git clone $prod_gitdb "$folderpath/sitebackups/proddb"
   fi
 
   if [[ -d "$site_path/$sitename_var" ]]; then
-     # Check that if the site exists, that it has the prod repo. Then only need to pull it!
-     ocmsg "The site: $sitename_var already exits. Now check if it is has the prod repo." debug
-     cd "$site_path/$sitename_var"
-     ocmsg "Local: $(git config --get remote.origin.url) Remote: $prod_gitrepo"
-     if [[ "$(git config --get remote.origin.url)" == "$prod_gitrepo" ]]; then
-       # Nice and simple!
-       ocmsg "Pull the files down." debug
-         git fetch --all
-        #git checkout -b backup-master
-        git reset --hard origin/master
-     else
-       ocmsg "Removing old $sitename_var site and cloning the files into $sitename_var" debug
-       # Set up the prod repo in the desired site location after deleting what is already there.
-       rm -rf "$site_path/$sitename_var"
-       cd $site_path
-       git clone $prod_gitrepo $sitename_var
-     fi
+    # Check that if the site exists, that it has the prod repo. Then only need to pull it!
+    ocmsg "The site: $sitename_var already exits. Now check if it is has the prod repo." debug
+    cd "$site_path/$sitename_var"
+    ocmsg "Local: $(git config --get remote.origin.url) Remote: $prod_gitrepo"
+    if [[ "$(git config --get remote.origin.url)" == "$prod_gitrepo" ]]; then
+      # Nice and simple!
+      ocmsg "Pull the files down." debug
+      git fetch --all
+      #git checkout -b backup-master
+      git reset --hard origin/master
+    else
+      ocmsg "Removing old $sitename_var site and cloning the files into $sitename_var" debug
+      # Set up the prod repo in the desired site location after deleting what is already there.
+      rm -rf "$site_path/$sitename_var"
+      cd $site_path
+      git clone $prod_gitrepo $sitename_var
+    fi
   else
     # clone the repo
     ocmsg "Cloning the files into $sitename_var" debug
@@ -202,42 +213,42 @@ if [[ "$bk" == "prod" ]] && [[ "$prod_method" == "git" ]] && [[ "$sitename_var" 
     git clone $prod_gitrepo $sitename_var
   fi
 
-    # now tar it so it is backed up for future use while we are at it.
+  # now tar it so it is backed up for future use while we are at it.
   #tar --exclude='$site_path/$sitename_var/$webroot/sites/default/settings.local.php' --exclude='$site_path/$sitename_var/$webroot/sites/default/settings.php' -zcf "$folderpath/sitebackups/prod/$bname.tar.gz" "$site_path/$sitename_var"
   fix_site_settings
 
-echo "Set site permissions"
-set_site_permissions $sitename_var
+  echo "Set site permissions"
+  set_site_permissions $sitename_var
 
-#restore db
-db_defaults
-echo -e "$Cyan Restore the database $Color_Off"
-restore_db
-echo -e "$Cyan Files and database have been restored $Color_Off"
-exit
-    else
-
-ocmsg "flag_first is $flag_first" debug
-options=($(find -maxdepth 1 -name "*.sql" -print0 | xargs -0 ls -1 -t))
-if [ $flag_first ]; then
-  echo -e "\e[34mrestoring $1 to $2 with latest backup\e[39m"
-  Name=${options[0]:2}
-  echo "Restoring with $Name"
+  #restore db
+  db_defaults
+  echo -e "$Cyan Restore the database $Color_Off"
+  restore_db
+  echo -e "$Cyan Files and database have been restored $Color_Off"
+  exit
 else
-  prompt="Please select a backup:"
-  PS3="$prompt "
-  select opt in "${options[@]}" "Quit"; do
-    if ((REPLY == 1 + ${#options[@]})); then
-      exit
-    elif ((REPLY > 0 && REPLY <= ${#options[@]})); then
-      echo "You picked $REPLY which is file ${opt:2}"
-      Name=${opt:2}
-      break
-    else
-      echo "Invalid option. Try another one."
-    fi
-  done
-fi
+
+  ocmsg "flag_first is $flag_first" debug
+  options=($(find -maxdepth 1 -name "*.sql" -print0 | xargs -0 ls -1 -t))
+  if [ $flag_first ]; then
+    echo -e "\e[34mrestoring $1 to $2 with latest backup\e[39m"
+    Name=${options[0]:2}
+    echo "Restoring with $Name"
+  else
+    prompt="Please select a backup:"
+    PS3="$prompt "
+    select opt in "${options[@]}" "Quit"; do
+      if ((REPLY == 1 + ${#options[@]})); then
+        exit
+      elif ((REPLY > 0 && REPLY <= ${#options[@]})); then
+        echo "You picked $REPLY which is file ${opt:2}"
+        Name=${opt:2}
+        break
+      else
+        echo "Invalid option. Try another one."
+      fi
+    done
+  fi
 fi
 echo " site_path: $site_path/$sitename_var"
 # Check to see if folder already exits.
@@ -259,16 +270,16 @@ echo -e "\e[34mrestoring files\e[39m"
 # Will need to first move the source folder ($bk) if it exists, so we can create the new folder $sitename_var
 echo "path $site_path/$sitename_var folderpath $folderpath"
 
-  mkdir "$site_path/$sitename_var"
-  echo "$folderpath/sitebackups/$bk/${Name::-4}.tar.gz into $site_path/$sitename_var"
-  # Check to see if the backup includes the root folder or not.
-  Dir_name=$(tar -tzf "$folderpath/sitebackups/$bk/${Name::-4}.tar.gz" | head -1 | cut -f1 -d"/")
-  #echo "Dir_name = >$Dir_name<"
-  if [ $Dir_name == "." ]; then
-    tar -zxf "$folderpath/sitebackups/$bk/${Name::-4}.tar.gz" -C "$site_path/$sitename_var"
-  else
-    tar -zxf "$folderpath/sitebackups/$bk/${Name::-4}.tar.gz" -C "$site_path/$sitename_var" --strip-components=1
-  fi
+mkdir "$site_path/$sitename_var"
+echo "$folderpath/sitebackups/$bk/${Name::-4}.tar.gz into $site_path/$sitename_var"
+# Check to see if the backup includes the root folder or not.
+Dir_name=$(tar -tzf "$folderpath/sitebackups/$bk/${Name::-4}.tar.gz" | head -1 | cut -f1 -d"/")
+#echo "Dir_name = >$Dir_name<"
+if [ $Dir_name == "." ]; then
+  tar -zxf "$folderpath/sitebackups/$bk/${Name::-4}.tar.gz" -C "$site_path/$sitename_var"
+else
+  tar -zxf "$folderpath/sitebackups/$bk/${Name::-4}.tar.gz" -C "$site_path/$sitename_var" --strip-components=1
+fi
 
 # Move settings.php and settings.local.php out the way before they are overwritten just in case you might need them.
 #echo "Moving settings.php and settings.local.php"
@@ -294,11 +305,10 @@ if [[ $flag_open ]]; then
   drush @$sitename_var uli &
 fi
 
-
 #drush @sitename_var cr
 # End timer
 ################################################################################
 # Finish script, display time taken
 ################################################################################
-echo 'Finished in H:'$(($SECONDS/3600))' M:'$(($SECONDS%3600/60))' S:'$(($SECONDS%60))
+echo 'Finished in H:'$(($SECONDS / 3600))' M:'$(($SECONDS % 3600 / 60))' S:'$(($SECONDS % 60))
 exit 0
